api/requirements.txt
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
arxiv==2.0.0
ase==3.22.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcbio-gff==0.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
biopython==1.81
black==23.9.1
blinker==1.6.2
boto3==1.28.60
botocore==1.31.60
build==1.0.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.0.7
cryptography==41.0.4
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
e3nn==0.4.4
#emmet-core==0.69.5
emmet-core==0.77.1
execnet==1.9.0
executing==1.2.0
fastapi==0.103.2
feedparser==6.0.10
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
greenlet==2.0.2
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.4
langchain==0.0.325
langchain-experimental==0.0.22
langchainhub==0.1.13
langsmith==0.0.56
latexcodec==2.0.1
lightning-utilities==0.9.0
llamaapi==0.1.36
mace @ git+https://github.com/chiang-yuan/mace.git@f8df28e8f3f018d4bf6b31ff37c3a25655db5064
maggma==0.57.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
matscipy==0.8.0
mongogrant==0.3.3
mongomock==4.1.2
monotonic==1.6
monty==2023.9.25
mp-api==0.37.2
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numexpr==2.8.6
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.16.1
openai==0.28.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
orjson==3.9.7
overrides==7.4.0
packaging==23.1
palettable==3.3.3
pandas==2.1.1
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip-tools==7.3.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.0.0
posthog==3.0.2
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pybtex==0.24.0
pycparser==2.21
#pydantic==2.4.2
pydantic==2.6.1
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydash==7.0.6
Pygments==2.16.1
pymatgen==2023.9.10
pymongo==4.5.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyNaCl==1.5.0
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==7.3.1
pytest-mock==3.10.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-hostlist==1.23.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.3
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
ruff==0.0.291
s3transfer==0.7.0
scipy==1.10.1
sentinels==1.0.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spglib==2.1.0
SQLAlchemy==2.0.21
sshtunnel==0.4.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torch-ema==0.3
torchmetrics==1.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
triton==2.1.0
typer==0.9.0
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uncertainties==3.1.7
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.6
websockets==12.0
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wikipedia==1.4.0
yarl==1.9.2
